+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.18.0] - November 13th 2020
+
+- New Features
+  - **common** - new utils `last`, `filterBy`, `uniqueBy`, and classes `OperationStack` and `OperationError`
+
 ## [6.17.0] - November 11th 2020
 
 ### Other Changes
@@ -188,13 +193,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### @esri/hub-content
 
 - Bug Fixes
-  - **normalizeItemType**: getType =&amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;gt; normalizeItemType, support for Hub Initiative Template types [`78a6fe0d`](https://github.com/Esri/hub.js/commit/78a6fe0d758e115c0b0ecc5f170ac913591153e2)
+  - **normalizeItemType**: getType =&amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;gt; normalizeItemType, support for Hub Initiative Template types [`78a6fe0d`](https://github.com/Esri/hub.js/commit/78a6fe0d758e115c0b0ecc5f170ac913591153e2)
   - **collections**: add template collection, getHubItemType only accepts items [`57b783a2`](https://github.com/Esri/hub.js/commit/57b783a242b7a2747db0dbc7fce57155c56ad93f) [#347](https://github.com/Esri/hub.js/issues/347) [#354](https://github.com/Esri/hub.js/issues/354)
 
 ### @esri/hub-sites
 
 - Bug Fixes
-  - **normalizeItemType**: getType =&amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;gt; normalizeItemType, support for Hub Initiative Template types [`78a6fe0d`](https://github.com/Esri/hub.js/commit/78a6fe0d758e115c0b0ecc5f170ac913591153e2)
+  - **normalizeItemType**: getType =&amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;amp;gt; normalizeItemType, support for Hub Initiative Template types [`78a6fe0d`](https://github.com/Esri/hub.js/commit/78a6fe0d758e115c0b0ecc5f170ac913591153e2)
 
 ### Other Changes
 
@@ -1238,4 +1243,5 @@ Initial Release
 [6.15.0]: https://github.com/Esri/hub.js/compare/v6.14.0...v6.15.0 "v6.15.0"
 [6.16.0]: https://github.com/Esri/hub.js/compare/v6.15.0...v6.16.0 "v6.16.0"
 [6.17.0]: https://github.com/Esri/hub.js/compare/v6.16.0...v6.17.0 "v6.17.0"
-[head]: https://github.com/Esri/hub.js/compare/v6.17.0...HEAD "Unreleased Changes"
+[6.18.0]: https://github.com/Esri/hub.js/compare/v6.17.0...v6.18.0 "v6.18.0"
+[head]: https://github.com/Esri/hub.js/compare/v6.18.0...HEAD "Unreleased Changes"